requirements.txt
opencv=4.1.0
pytorch